# openapi.yaml
openapi: 3.1.0
info:
  title: RTK Billing API
  version: 1.0.0
  description: Provider-neutral pricing, wallet, payment, invoice, and billing access
    service.
paths:
  /v1/internal/billing/pricing-versions:
    post:
      tags:
      - Internal Billing
      summary: Create a draft pricing version and immutable rate set
      operationId: createBillingPricingVersion
      security:
      - billingServiceAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '201':
          description: Pricing version created.
        '409':
          $ref: '#/components/responses/Error'
  /v1/internal/billing/pricing-versions/{pricingVersionId}/activate:
    post:
      tags:
      - Internal Billing
      summary: Activate one pricing version and retire the prior version
      operationId: activateBillingPricingVersion
      security:
      - billingServiceAuth: []
      parameters:
      - name: pricingVersionId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Pricing version activated.
        '404':
          $ref: '#/components/responses/Error'
  /v1/internal/billing/usage-facts:
    post:
      tags:
      - Internal Billing
      summary: Idempotently ingest a digest-anchored usage fact
      operationId: putBillingUsageFact
      security:
      - billingServiceAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Existing identical fact.
        '201':
          description: Usage fact created.
        '409':
          $ref: '#/components/responses/Error'
  /v1/internal/billing/periods/close:
    post:
      tags:
      - Internal Billing
      summary: Close, price, issue, settle, and render one billing period
      operationId: closeBillingPeriod
      security:
      - billingServiceAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Existing or newly closed invoice result.
        '422':
          $ref: '#/components/responses/Error'
  /v1/orgs/{orgId}/billing/summary:
    get:
      tags:
      - Billing and Payments
      summary: Return tenant balance current cost runway and top-up health
      operationId: getBillingSummary
      security:
      - billingServiceAuth: []
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Tenant billing summary.
        '403':
          $ref: '#/components/responses/Error'
  /v1/orgs/{orgId}/billing/usage:
    get:
      tags:
      - Billing and Payments
      summary: Return server-priced usage for the active billing period
      operationId: getBillingUsage
      security:
      - billingServiceAuth: []
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Priced usage lines.
        '403':
          $ref: '#/components/responses/Error'
  /v1/orgs/{orgId}/billing/invoices:
    get:
      tags:
      - Billing and Payments
      summary: List tenant invoices with bounded filters
      operationId: listBillingInvoices
      security:
      - billingServiceAuth: []
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Invoice page.
        '403':
          $ref: '#/components/responses/Error'
  /v1/orgs/{orgId}/billing/invoices/{invoiceId}:
    get:
      tags:
      - Billing and Payments
      summary: Get one tenant invoice and settlement links
      operationId: getBillingInvoice
      security:
      - billingServiceAuth: []
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
      - name: invoiceId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Invoice detail.
        '404':
          $ref: '#/components/responses/Error'
  /v1/orgs/{orgId}/billing/invoices/{invoiceId}/pdf:
    get:
      tags:
      - Billing and Payments
      summary: Download an immutable issued invoice PDF
      operationId: downloadBillingInvoicePdf
      security:
      - billingServiceAuth: []
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
      - name: invoiceId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Immutable invoice document
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/Error'
  /v1/orgs/{orgId}/billing/activity:
    get:
      tags:
      - Billing and Payments
      summary: List normalized customer-safe billing activity
      operationId: listBillingActivity
      security:
      - billingServiceAuth: []
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Billing activity page.
        '403':
          $ref: '#/components/responses/Error'
  /v1/orgs/{orgId}/billing/activity/{activityId}:
    get:
      tags:
      - Billing and Payments
      summary: Get one customer-safe billing activity timeline
      operationId: getBillingActivity
      security:
      - billingServiceAuth: []
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
      - name: activityId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Billing activity detail.
        '404':
          $ref: '#/components/responses/Error'
  /v1/orgs/{orgId}/billing/profile:
    get:
      tags:
      - Billing and Payments
      summary: Get the prospective billing recipient profile
      operationId: getBillingProfile
      security:
      - billingServiceAuth: []
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Billing profile.
    put:
      tags:
      - Billing and Payments
      summary: Replace the prospective billing recipient profile
      operationId: putBillingProfile
      security:
      - billingServiceAuth: []
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
      - name: If-Match
        in: header
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Updated billing profile.
        '412':
          $ref: '#/components/responses/Error'
  /v1/orgs/{orgId}/billing/statements:
    get:
      tags:
      - Billing and Payments
      summary: Export tenant billing activity as UTF-8 CSV
      operationId: exportBillingStatement
      security:
      - billingServiceAuth: []
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: CSV statement
          content:
            text/csv:
              schema:
                type: string
...
  /v1/orgs/{orgId}/auto-topup:
    get:
      tags:
      - Billing and Payments
      operationId: getAutoTopUpPolicy
      x-rtk-feature-id: FEAT-AM-PAYMENT-001
      x-rtk-requirement-ids:
      - REQ-AM-AUTO-TOPUP-001
      - REQ-AM-PAYMENT-AUTHZ-001
      security:
      - billingServiceAuth: []
      summary: Read the automatic top-up policy and Asia/Taipei safety-limit reset.
      parameters:
      - $ref: '#/components/parameters/OrgId'
      responses:
        '200':
          description: Policy response; auto_topup is null when no policy exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutoTopUpResponse'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
    put:
      tags:
      - Billing and Payments
      operationId: replaceAutoTopUpPolicy
      x-rtk-feature-id: FEAT-AM-PAYMENT-001
      x-rtk-requirement-ids:
      - REQ-AM-AUTO-TOPUP-001
      - REQ-AM-PAYMENT-METHOD-001
      - REQ-AM-PAYMENT-AUDIT-001
      security:
      - billingServiceAuth: []
      summary: Replace the automatic top-up policy using optimistic concurrency and
        explicit consent.
      parameters:
      - $ref: '#/components/parameters/OrgId'
      - $ref: '#/components/parameters/IfMatchVersion'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PutAutoTopUpRequest'
      responses:
        '200':
          description: Policy replaced with a new generation and version.
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutoTopUpResponse'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '428':
          $ref: '#/components/responses/Error'
    delete:
      tags:
      - Billing and Payments
      operationId: disableAutoTopUpPolicy
      x-rtk-feature-id: FEAT-AM-PAYMENT-001
      x-rtk-requirement-ids:
      - REQ-AM-AUTO-TOPUP-001
      - REQ-AM-PAYMENT-AUDIT-001
      security:
      - billingServiceAuth: []
      summary: Disable automatic top-up and advance the policy generation.
      parameters:
      - $ref: '#/components/parameters/OrgId'
      - $ref: '#/components/parameters/IfMatchVersion'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentMutationReasonRequest'
      responses:
        '200':
          description: Disabled policy.
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutoTopUpResponse'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '428':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    billingServiceAuth:
      type: http
      scheme: bearer
      bearerFormat: service-token
      description: Dedicated Cloud Admin or trusted billing-worker service credential.
  schemas:
...
    AutoTopUpResponse:
      type: object
      required:
      - auto_topup
      properties:
        auto_topup:
          allOf:
          - $ref: '#/components/schemas/AutoTopUpPolicy'
          nullable: true
...
    PaymentMutationReasonRequest:
      type: object
      additionalProperties: false
      required:
      - reason
      properties:
        reason:
          type: string
          minLength: 3
          maxLength: 500
...
    PutAutoTopUpRequest:
      type: object
      additionalProperties: false
      required:
      - enabled
      - threshold_minor
      - top_up_amount_minor
      - currency
      - payment_method_id
      - daily_attempt_limit
      - daily_amount_limit_minor
      - cooldown_seconds
      - consent
      properties:
        enabled:
          type: boolean
        threshold_minor:
          $ref: '#/components/schemas/MoneyMinor'
        top_up_amount_minor:
          $ref: '#/components/schemas/MoneyMinor'
        currency:
          $ref: '#/components/schemas/Currency'
        payment_method_id:
          type: string
          format: uuid
        daily_attempt_limit:
          type: integer
          minimum: 1
          maximum: 10
        daily_amount_limit_minor:
          $ref: '#/components/schemas/MoneyMinor'
        cooldown_seconds:
          type: integer
          format: int64
          minimum: 300
        consent:
          $ref: '#/components/schemas/PaymentConsentAcceptance'
...
    MoneyMinor:
      type: integer
      format: int64
      minimum: 1
      description: Currency-unit integer. For zero-decimal TWD, one amount_minor unit
        is NT$1.
...
  parameters:
...
    OrgId:
      name: orgId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    IfMatchVersion:
      name: If-Match
      in: header
      required: true
      description: Quoted or unquoted numeric automatic top-up policy version; use
        0 to create the first policy.
      schema:
        type: string
        pattern: ^"?[0-9]+"?$
...
  responses:
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalid_claim_token:
              summary: Non-retryable Claim Token error.
              value:
                error:
                  code: invalid_claim_token
                  message: Invalid claim token
                  retryable: false
                  resolution_action: scan_or_enter_a_valid_claim_token
            claim_service_unavailable:
              summary: Retryable Claim Token service error.
              value:
                error:
                  code: service_unavailable
                  message: Claim resolve is temporarily unavailable
                  retryable: true
                  resolution_action: retry_later
            oidc_user_not_provisioned:
              summary: Unknown SSO user.
              value:
                error:
                  code: user_not_provisioned
                  message: SSO user is not provisioned in account manager
            oidc_disabled:
              summary: OIDC login disabled.
              value:
                error:
                  code: oidc_disabled
                  message: OIDC authentication is disabled
            invalid_oidc_state:
              summary: Invalid or replayed OIDC state.
              value:
                error:
                  code: invalid_oidc_state
                  message: Invalid OIDC login state
            invalid_oidc_token:
              summary: Invalid OIDC token.
              value:
                error:
                  code: invalid_oidc_token
                  message: Invalid OIDC token
            unverified_oidc_email:
              summary: Unverified OIDC email.
              value:
                error:
                  code: unverified_oidc_email
                  message: OIDC email is not verified
            oidc_provider_misconfigured:
              summary: OIDC provider misconfigured.
              value:
                error:
                  code: oidc_provider_misconfigured
                  message: OIDC provider is misconfigured